+++ plugins/core/span_noop.go
@@ -25,6 +25,7 @@ const noopContextValue = "N/A"
 
 type NoopSpan struct {
 	stackCount int
+	tracer     *Tracer
 }
 
 func newSnapshotNoopSpan() *NoopSpan {
@@ -34,9 +35,10 @@ func newSnapshotNoopSpan() *NoopSpan {
 	}
 }
 
-func newNoopSpan() *NoopSpan {
+func newNoopSpan(tracer *Tracer) *NoopSpan {
 	return &NoopSpan{
 		stackCount: 1,
+		tracer:     tracer,
 	}
 }
 
@@ -89,6 +91,9 @@ func (*NoopSpan) Log(...string) {
 func (*NoopSpan) Error(...string) {
 }
 
+func (*NoopSpan) ErrorOccured() {
+}
+
 func (n *NoopSpan) enterNoSpan() {
 	n.stackCount++
 }
@@ -96,6 +101,7 @@ func (n *NoopSpan) enterNoSpan() {
 func (n *NoopSpan) End() {
 	n.stackCount--
 	if n.stackCount == 0 {
+		GetSo11y(n.tracer).MeasureTracingContextCompletion(true)
 		if ctx := getTracingContext(); ctx != nil {
 			ctx.SaveActiveSpan(nil)
 		}