+++ tools/go-agent/instrument/plugins/rewrite/context.go
@@ -45,6 +45,7 @@ var OperatePrefix = GenerateCommonPrefix + "operator"
 var TypePrefix = OperatePrefix + "Type"
 var VarPrefix = OperatePrefix + "Var"
 var StaticMethodPrefix = OperatePrefix + "StaticMethod"
+var BeforeInterStart = "beforeInterStart"
 
 type Context struct {
 	pkgFullPath   string
@@ -192,7 +193,9 @@ func (c *Context) enhanceTypeNameWhenRewrite(fieldType dst.Expr, parent dst.Node
 		if _, ok := parent.(*dst.CallExpr); ok {
 			t.Name = fmt.Sprintf("%s%s%s", StaticMethodPrefix, c.currentPackageTitle, name)
 		} else {
-			t.Name = fmt.Sprintf("%s%s%s", TypePrefix, c.currentPackageTitle, name)
+			if !strings.Contains(name, BeforeInterStart) {
+				t.Name = fmt.Sprintf("%s%s%s", TypePrefix, c.currentPackageTitle, name)
+			}
 		}
 		return name, t.Name
 	case *dst.SelectorExpr:
@@ -341,7 +344,7 @@ func (c *Context) enhanceTypeNameWhenRewrite(fieldType dst.Expr, parent dst.Node
 
 func (c *Context) typeIsBasicTypeValueOrEnhanceName(name string) bool {
 	if strings.HasPrefix(name, OperatePrefix) || strings.HasPrefix(name, GenerateMethodPrefix) || tools.IsBasicDataType(name) ||
-		name == "nil" || name == "true" || name == "false" || name == "append" || name == "panic" || name == "new" {
+		name == "nil" || name == "true" || name == "false" || name == "append" || name == "panic" || name == "new" || name == "delete" {
 		return true
 	}
 	if _, valErr := strconv.ParseFloat(name, 64); valErr == nil {