+++ CHANGES.md
@@ -9,6 +9,7 @@ Release Notes.
 * support attaching events to span in the toolkit.
 * support record log in the toolkit.
 * support manually report metrics in the toolkit.
+* support manually set span error in the toolkit.
 
 #### Plugins
 * Support [goframev2](https://github.com/gogf/gf) goframev2.
@@ -21,6 +22,9 @@ Release Notes.
 #### Bug Fixes
 * Fix wrong docker image name and `-version` command.
 * Fix redis plugin cannot work in cluster mode.
+* Fix cannot find file when exec build in test/plugins.
+* Fix not set span error when http status code >= 400
+* Fix http plugin cannot provide peer name when optional Host is empty.
 
 #### Issues and PR
 - All issues are [here](https://github.com/apache/skywalking/milestone/219?closed=1)